% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_all_lessons.R
\name{build_all_lessons}
\alias{build_all_lessons}
\title{Build all lessons for the package.}
\usage{
build_all_lessons(
  source_dir = file.path(getwd(), "source"),
  learnr = TRUE,
  github = TRUE
)
}
\arguments{
\item{source_dir}{The base directory containing the tutorial sub directories.}

\item{learnr}{Logical, if TRUE, builds LearnR tutorials from the YAML files.}

\item{github}{Logical, if TRUE, builds GitHub markdown files from the YAML files.}
}
\value{
A list containing the results of processing each YAML file.
}
\description{
This function reads the YAML source for each lesson and builds the training materials for the package.
Depending on the options specified, it can build LearnR tutorials or GitHub markdown files.
}
