% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_learnr.R
\name{build_exercise}
\alias{build_exercise}
\title{Build individual exercise content for a learnr tutorial}
\usage{
build_exercise(exercise, exercise_number)
}
\arguments{
\item{exercise}{A list containing details of the exercise, such as instructions, hints,
and solutions.}

\item{exercise_number}{An integer indicating the exercise sequence number which is used
to uniquely identify exercise chunks.}
}
\value{
A character string containing the complete R Markdown for an exercise, including
instructions, hints, solution code, and checking code.
}
\description{
This function takes a single exercise object and formats it into R Markdown. It handles
the creation of exercise prompts, hints, solutions, and the corresponding checking code
using the \code{gradethis} package. The output is designed to integrate seamlessly into a
learnr tutorial.
}
