% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_learnr.R
\name{build_metadata}
\alias{build_metadata}
\title{Build metadata for a learnr tutorial}
\usage{
build_metadata(metadata, type)
}
\arguments{
\item{metadata}{A list of existing metadata from the tutorial template.}

\item{type}{A character string specifying the type of tutorial, supports only "learnr".}
}
\value{
A character string with YAML formatted metadata, wrapped in appropriate YAML document delimiters.
}
\description{
Constructs a list of metadata settings specific to learnr tutorials, merging it with existing
metadata from a template and formatting it into YAML.
}
