% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_learnr.R
\name{build_next_lesson}
\alias{build_next_lesson}
\title{Build the navigation for the next lesson in a learnr tutorial}
\usage{
build_next_lesson(lesson)
}
\arguments{
\item{lesson}{A list containing details of the lesson, particularly the lesson number.}
}
\value{
A character string with R markdown format for UI and server blocks in a learnr tutorial.
}
\description{
Generates markdown formatted text for a button to mark the current lesson as complete and
proceed to the next lesson. Includes both UI and server side reactive expressions for Shiny.
}
