% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_github_md.R
\name{generate_toc}
\alias{generate_toc}
\title{Generate a Table of Contents for GitHub Markdown}
\usage{
generate_toc(content, depth = 1)
}
\arguments{
\item{content}{The content structure from the YAML template}
}
\value{
A Markdown formatted table of contents
}
\description{
Generate a Table of Contents for GitHub Markdown
}
