% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_learnr.R
\name{generateCodeBlockName}
\alias{generateCodeBlockName}
\title{Generate a random name for a code block}
\usage{
generateCodeBlockName()
}
\value{
A character string containing a randomly generated code block name.
}
\description{
This function creates a random string that can be used as a name for R code chunks.
It's useful for dynamically generating unique names for code chunks in markdown.
}
