% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userstate.R
\name{get_user_state}
\alias{get_user_state}
\title{Retrieve a parameter from the user state}
\usage{
get_user_state(name)
}
\arguments{
\item{name}{The name of the parameter to retrieve from the user state.}
}
\value{
The value of the specified parameter, or \code{NULL} if it does not exist.
}
\description{
Accesses a specific parameter from the user state by name. If the parameter is not
found, it returns \code{NULL}.
}
\keyword{internal}
