% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_learnr.R
\name{process_exercises}
\alias{process_exercises}
\title{Process exercises for a learnr tutorial}
\usage{
process_exercises(exercise_objects)
}
\arguments{
\item{exercise_objects}{A list of lists, where each sub-list contains elements that define an exercise.}
}
\value{
A character string containing all exercises formatted in R markdown.
}
\description{
Converts a list of exercise specifications into formatted R markdown for interactive exercises
in a learnr tutorial. Each exercise includes instructions, hints, solutions, and checking logic.
}
