% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lesson.R
\name{run_lesson}
\alias{run_lesson}
\title{Run a specified lesson from the available tutorials}
\usage{
run_lesson(number = NA)
}
\arguments{
\item{number}{An optional integer specifying the lesson number to run.
If NA (default), the function will prompt the user to choose a lesson.}
}
\value{
NULL, the function is used for its side effects of running a lesson interactively.
}
\description{
This function allows the user to run a specific lesson from the available tutorials in the \code{trainingRIntro} package.
It prompts the user to select a lesson if not specified. It checks the validity of the input and handles errors
appropriately, ensuring that only available lessons can be run.
}
\details{
If \code{number} is not provided or is NA, the function prompts the user interactively to choose a lesson from
a list of available tutorials. The function ensures that the input corresponds to a valid lesson number
and provides error feedback for invalid entries.
}
