% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userstate.R
\name{save_user_state}
\alias{save_user_state}
\title{Save the user state to disk}
\usage{
save_user_state(user_state)
}
\arguments{
\item{user_state}{A list representing the user state to be saved.}
}
\description{
Persists the current state of the user, such as lesson progress and preferences, to a file
on disk.
}
\keyword{internal}
