% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userstate.R
\name{set_user_state}
\alias{set_user_state}
\title{Set parameters in the user state}
\usage{
set_user_state(...)
}
\arguments{
\item{...}{Named arguments where each name is the user state variable to update, and
the corresponding value is the new value for that variable.}
}
\description{
Updates specific parameters within the user state. It first loads the current state,
modifies it based on provided arguments, and saves the updated state back to disk.
}
\keyword{internal}
