% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userstate.R
\name{user_state_file}
\alias{user_state_file}
\title{Determine the file path for storing user state}
\usage{
user_state_file()
}
\value{
A string containing the path to the file where user state is or will be stored.
}
\description{
This function computes the path to the file used for storing user state information.
It ensures the directory exists, creating it if necessary.
}
\keyword{internal}
